% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{core_samples}
\alias{core_samples}
\title{Reference sample annotations}
\format{
A data frame.
}
\usage{
core_samples
}
\description{
Packaged sample annotation data frame for the reference/core cohort.
Row names match \code{colnames(exp_core_g)} and the data frame contains a \code{CTS}
column.
}
\keyword{datasets}
