% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{exp_core_g}
\alias{exp_core_g}
\title{Reference gene expression matrix}
\format{
A numeric matrix.
}
\usage{
exp_core_g
}
\description{
Packaged reference/core expression matrix with genes in rows and samples
in columns.
}
\keyword{datasets}
